Refactor import paths for RPS and Solvedoku windows
- Updated import statements in rock-paper-scissors.tsx and solvedoku.tsx to reflect new directory structure for better organization.
- Ensured that the components are correctly imported from their respective paths, improving code maintainability.
